Setting up initial code changes for Team and Player, Match and Tournament Entities along with corresponding Controller, Service and Repository layers.
This commit sets up the code to implement the most basic first versions of the entities that of Match, Player, Tournament, Venue and Team to form the backbone of the backend of the Sports Tournament Organiser Project. For a three Controllers, Services and Repositories have been implemented for those entities, enabling the most basic CRUD operations to be carried out on those entities. What needs to be noted is that the implementation of creation of many of the entities is only in its most elementary form as for now and relationships between entities such as ManyToOne, OneToMan are not yet fully implemented and are to be added and expanded on in a future commit, following this one. The Participation entity with its corresponding relationships and its accompanying Controller, Service And Repository layers is to be added in a future work. 

---------

Co-authored-by: Alexander <alexander.karaivanov@mentormate.com>
Co-authored-by: Aleksandar <conevc@uni-sofia.bg>