Initial changes for adding the Participation entity with Controller, Service And Repository Layers.
This commit contains the initial changes needed to introduce the Participation entity with its corresponding Controller, Service And Repository Layers and to set up the relationship between the Participation and Tournament entities, as well as updates on the logic of creating and updating the fields of other entities, such as Match, Player and Team. The commit also includes changes on includes mappers for creating Data Transfer Objects for Match, Participation and Tournament. 


---------

Co-authored-by: Alexander <alexander.karaivanov@mentormate.com>
Co-authored-by: Stefcho1227 <stefan.ivanov1227@gmail.com>