Introduce changes needed to work more efficiently with the Player and Team entities in the application with the help of Input Data Transfer Objects and Output Data Transfer Objects for Player and Team. Include mappers for both entities. Also introducing changes to the Controllers and Services of those entities to accomodate the newly added Data Transfer Objects for those entties and their accompanying Mappers. (#4)
In order to improve the smoothness of operations over the Player entity and the Team entity, this commit makes additions to the codebase to add Input Data Transfer Objects and Output Data Transfer Objects for Player and Team. Along with them we also create mappers, which enable conversions from Input DTO to an Entity, from an Entity to DTO, and it also allows for updating a given currently existing entity. In this way we can more easily manage the bidirectional relationship between Players and Teams. This also enables us to easily track changes of any occurring changes within Teams and the Players by inspecting the other entity via the usage of the bidirectional mapping.