Introduce first version of frontend with backend entities
This commit introduces the first version of the front end for the Sports Tournament Organiser based on Reac. The changes introduced here allow the creation and of players and teams, adding players to teams, editing player information. The commit also makes changes to the back end to allow the front end and the backend to work together. This commit also makes changes to the JWT authentication mechanism.